% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assessment_gen.R
\name{assessment_gen}
\alias{assessment_gen}
\title{Assessment Data Generation}
\usage{
assessment_gen(correlation_matrix, n = 100, means, sd, effect_size)
}
\arguments{
\item{correlation_matrix}{A square matrix specifying the desired correlations between variables.}

\item{n}{The number of observations to generate.}

\item{means}{A numeric vector specifying the mean for each variable.}

\item{sd}{A numeric vector specifying the standard deviation for each variable.}

\item{effect_size}{A numeric vector of effect sizes to simulate adverse impacts between groups.}
}
\value{
A dataframe of simulated data with columns corresponding to variables and a 'Race' column.
}
\description{
Assessment Data Generation
}
\examples{
Omega <- as.data.frame(matrix(c(1, 0.8, 0.3, 0.8, 1, 0.8, 0.3, 0.8, 1), byrow = TRUE, ncol = 3))
name <- c("GMA", "Personality", "Performance")
colnames(Omega) <- name
means <- c(5, 5, 5)
sd <- c(1, 1, 1)
effect_size <- c(1.5, 0.3, 0)
simulated_data <- assessment_gen(Omega, n = 1000, means = means, sd = sd, effect_size)

}
