% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songPareto.R
\name{songPareto}
\alias{songPareto}
\title{Chelsea Song's Pareto-optimization Procedure}
\usage{
songPareto(
  data,
  predictors,
  sr,
  d = NULL,
  corrMat = NULL,
  Spac = 20,
  graph = T,
  display_solution = T
)
}
\arguments{
\item{data}{Dataframe generated via AssessmentGen methods.}

\item{predictors}{Character vector containing strings corresponding to predictors of interest}

\item{sr}{A priori selection ratio.}

\item{d}{Numeric vector of subgroup difference magnitudes. If no vector is provided,
B-W differences from package will be used.}

\item{corrMat}{Correlation matrix containing relationships between each predictor.
If no matrix is provided, correlation matrix from package will be used.}

\item{Spac}{Number of solutions devised via Pareto-optimization.}

\item{graph}{Logical; should graph be shown?}

\item{display_solution}{Logical; should solution be shown?}
}
\value{
Pareto-optimization graphs and solutions.
}
\description{
Chelsea Song's Pareto-optimization Procedure
}
\examples{
data <- auto_generation(n = 1000, seed = 42)
          songPareto(data, c("Structured Interview", "GMA tests", "Conscientiousness"), sr = .20)
}
