% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_generation.R
\name{auto_generation}
\alias{auto_generation}
\title{Assessment Data Generation}
\usage{
auto_generation(n = 100)
}
\arguments{
\item{n}{The number of observations to generate.}
}
\value{
A dataframe of simulated data with columns corresponding to variables and a 'Race' column.
}
\description{
Assessment Data Generation
}
\examples{
simulated_data <- auto_generation(n = 100)
}
